feat(store): rename media to IeObjects
