+++ src/modules/account/hooks/get-folders.ts
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import { foldersService } from '@account/services/folders';
 import { GetFoldersResponse } from '@account/types';
 import { QUERY_KEYS } from '@shared/const/query-keys';
-import { setFolders } from '@shared/store/media';
+import { setFolders } from '@shared/store/ie-objects';
 
 export function useGetFolders(enabled = true): UseQueryResult<GetFoldersResponse> {
 	const dispatch = useDispatch();
