+++ src/modules/shared/store/ie-objects/ie-objects.select.ts
@@ -5,9 +5,10 @@ import { AppState } from '../store.types';
 
 import { IeObjectSearchAggregations } from 'modules/ie-objects/types';
 
-export const selectMediaResults = (state: AppState): GetIeObjectsResponse | undefined =>
-	state.media.results;
-export const selectMediaFilterOptions = (state: AppState): IeObjectSearchAggregations | undefined =>
-	state.media.filterOptions;
+export const selectIeObjectsResults = (state: AppState): GetIeObjectsResponse | undefined =>
+	state.IeObjects.results;
+export const selectIeObjectsFilterOptions = (
+	state: AppState
+): IeObjectSearchAggregations | undefined => state.IeObjects.filterOptions;
 export const selectFolders = (state: AppState): GetFoldersResponse | undefined =>
-	state.media.folders;
+	state.IeObjects.folders;
