+++ src/modules/shared/store/ie-objects/index.ts
@@ -1,0 +1,3 @@
+export * from './ie-objects.select';
+export * from './ie-objects.slice';
+export * from './ie-objects.types';
