--- src/modules/shared/store/media/index.ts
@@ -1,3 +1,0 @@
-export * from './media.select';
-export * from './media.slice';
-export * from './media.types';
