+++ src/modules/visitor-space/components/GenreFilterForm/GenreFilterForm.tsx
@@ -9,7 +9,7 @@ import { useQueryParams } from 'use-query-params';
 import { SearchBar } from '@shared/components';
 import { CheckboxList } from '@shared/components/CheckboxList';
 import useTranslation from '@shared/hooks/use-translation/use-translation';
-import { selectMediaFilterOptions } from '@shared/store/media';
+import { selectIeObjectsFilterOptions } from '@shared/store/ie-objects';
 import { visitorSpaceLabelKeys } from '@visitor-space/const';
 import { VisitorSpaceFilterId } from '@visitor-space/types';
 
@@ -37,7 +37,7 @@ const GenreFilterForm: FC<GenreFilterFormProps> = ({ children, className }) => {
 		defaultValues,
 	});
 
-	const buckets = (useSelector(selectMediaFilterOptions)?.schema_genre.buckets || []).filter(
+	const buckets = (useSelector(selectIeObjectsFilterOptions)?.schema_genre.buckets || []).filter(
 		(bucket) => bucket.key.toLowerCase().includes(search.toLowerCase())
 	);
